requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
alembic==1.15.2
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
croniter==6.0.0
cryptography==44.0.2
debtcollector==3.0.0
decorator==5.2.1
dnspython==2.7.0
dogpile.cache==1.3.4
eval_type_backport==0.2.2
eventlet==0.39.1
faiss-cpu==1.10.0
fastapi==0.115.12
fasteners==0.19
frozenlist==1.5.0
futurist==3.1.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iso8601==2.1.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keystoneauth1==5.10.0
keystonemiddleware==10.9.0
kombu==5.5.2
legacy-cgi==2.6.3
Mako==1.3.9
MarkupSafe==3.0.2
mistral-lib==3.3.1
mistralai==0.4.2
msgpack==1.1.0
multidict==6.3.2
netaddr==1.3.0
networkx==3.4.2
numpy==2.2.4
orjson==3.10.16
os-service-types==1.7.0
oslo.cache==3.10.1
oslo.concurrency==7.1.0
oslo.config==9.7.1
oslo.context==5.7.1
oslo.db==17.2.1
oslo.i18n==6.5.1
oslo.log==7.1.0
oslo.messaging==16.1.0
oslo.metrics==0.11.0
oslo.middleware==6.3.1
oslo.policy==4.5.1
oslo.serialization==5.7.0
oslo.service==4.1.1
oslo.utils==8.2.0
osprofiler==4.2.0
packaging==24.2
paramiko==3.5.1
Paste==3.10.1
PasteDeploy==3.1.0
pbr==6.1.1
pecan==1.6.0
ply==3.11
prettytable==3.16.0
prometheus_client==0.21.1
propcache==0.3.1
psutil==7.0.0
pycadf==4.0.1
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
PyMuPDF==1.25.5
PyNaCl==1.5.0
pyparsing==3.2.3
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-keystoneclient==5.6.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
repoze.lru==0.7
requests==2.32.3
rfc3986==2.0.0
Routes==2.5.1
rpds-py==0.24.0
setuptools==78.1.0
simplegeneric==0.8.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
statsd==4.0.1
stevedore==5.4.1
tenacity==9.1.2
testresources==2.0.1
testscenarios==0.5.0
testtools==2.7.2
tooz==6.3.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
WebOb==1.8.9
wrapt==1.17.2
WSME==0.12.1
yappi==1.6.10
yaql==3.0.0
yarl==1.18.3
zipp==3.21.0
